Add USB-PD 4b5b encoder skeleton, tests.
The 4b5b encoder translates raw data (or control signals) to a form
that can be transmitted on the line. Four input bits become five output
bits where the value of the output is selected to facilitate both DC
balance and the presence of enough transitions for a receiver to lock
to whatever input it receives.