Add BMC transmitter control unit skeleton.
The intention is that the control unit will eventually be responsible
for receiving requests from a Wishbone bus, translating them into the
signals to operate the various transmitter blocks, and processing any
signals received from those blocks as necessary for a transmission.

The full skeleton is not provided here, only the Wishbone signals. It
may be the case that it makes more sense to divide some functionality
out into separate blocks. More thought will be put towards this after
a functional Wishbone interface has been attained.