Reorganise files for control unit development.
Design and test files are renamed to better indicate the purpose of the
units they contain and to reflect a change of purpose.

Ideally, it would be possible to produce automated ('self-testing') test
cases for the BMC transmitter control unit. However, to avoid cluttering
the control unit's port definition, it would be necessary to alias any
internal signals required. Specifically, it was intended this would be
done to check internal register values without completing a Wishbone bus
transaction. Unfortunately, GHDL does not support this kind of aliases
and does not appear to intend to support them, and Vivado appears not to
support VHDL-2008 to the extent required to use VUnit generally.

Resolving this, 'BiphaseMarkCU' becomes 'BiphaseMarkTransmitter' and
will now contain the other components, glue logic, Wishbone logic, and
anything else required. This is less than ideal as it produces a very
large single unit, but appears necessary. Automated testing can then be
carried out at the 'acceptance' or 'integration' test level.

There will also be some issue in simulation as it is intended that hard
RAM/FIFO blocks on the FPGA will be used internally. It may be viable to
create a (non-synthesisable) mock entity for use in testing.