Add FIFO9 block skeleton.
It is intended that FIFO9 will provide a common interface for both a
hardware block RAM/FIFO primitive and a FIFO implemented in the FPGA
fabric. This way, it should hopefully be possible to simulate and
automatically test the BMC transmitter.

A minor correction is also made to a comment on the BMC transmitter.