Correct BMC transmitter write test.
The previous test distinguished between lower and higher data halves by
testing for every 5th and every 10th cycle. This would be viable if all
data came in two halves, but K-codes don't. Thus, a K-code followed by
raw data would see a K-code on the Cycle 5, the lower half on Cycle 10,
and the higher half on Cycle 15. The higher half wouldn't be correctly
detected in the previous implementation as 15 mod 10 /= 0.