Fix false-positive error in RX overlong detection.
In the 'slow_bmc' test (i.e. at the very slowest frequency permitted
by USB-PD), monitoring the previous 12us for edges would cause line
idle to be detected very briefly, resulting in the receiver entering
an error state for an unanticipated idle condition.

As this condition was very brief (single-digit WB cycles), the easiest
solution is to extend the monitor period so that more edges will be
taken into account. It was thought extending to 13us would be enough,
and so (to give a safety margin) the buffer is extended to 14us.

Minor consequential changes to tests are also made.