Add FIFO9 mapping to Xilinx BRAM.
The Artix-7 targeted in prototyping includes block RAMs (BRAMs) that
can be configured as either synchronous or asynchronous FIFOs. This
commit adds a mapping to the BRAM primitive, which should allow better
resource-use estimation from synthesis/PNR as the FIFO will no longer
be implemented in the fabric.

It is entirely untested.